+++ app/Http/Controllers/Admin/CustomerController.php
@@ -299,7 +299,7 @@ class CustomerController extends AdminController
             if ($user->id !== Auth::user()->id) {
                 try {
                     if ($user->delete()) {
-                        if($detail->delete()) {
+                        if($detail && $detail->delete()) {
                             if (Storage::disk('avatar')->exists($detail->avatar)) {
                                 if (strpos($detail->avatar, "default/") !== 0) {
                                     Storage::delete('public/avatars/' . $detail->avatar);
